requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Werkzeug==2.0.1
pg8000==1.29.4
gunicorn==20.1.0
python-dotenv==0.19.0
weasyprint==54.2